Updated to include price and tweet data
First attempt at committing changes.

I did quite a bit of work to try and relate tweets to stock price.  I also experimented with sentiment and stock price return for 1, 10, and 30 days (see last part of code), but not sure this worked the way I wanted it to.  Help would be great!