Last update to clean up my code
This should be my final EDA code.  Cleaned it up to make it neater, added figure saving, improved a few pieces to give good EDA results.  I have had this open on my computer for a few days, so I did not capture some of the other changes Aniruddha made to the last pull request to the master.  Sorry!!!